refactor: move schema-DDL recognition to statement position, split index module
Two code-health findings from the source-quality gate, fixed rather than
baselined around.

`parse_clause_sequence` had picked up two guarded arms and crossed into
"complex function". A schema command is a whole statement, not a pipeline
stage, so recognition moves to `parse_query`: one peek-only check per
query instead of one per clause, the clause loop keeps its original
shape, and the standalone-statement rule is now structural rather than
asserted. A misplaced `CREATE INDEX` — after another clause, or inside a
`CALL { }` body — is named by `parse_create_clause`, which is the only
place that can still see one.

`dir_graph/mod.rs` crossed the 2,500-line god-file ceiling. Split the
hash, composite, and B-tree index stores plus their incremental
maintenance into `dir_graph/indexes.rs` (577 lines), a self-contained
concern with one entry point per index kind; mod.rs drops to 1,941. The
GraphBackend-match whitelist follows the code.

The two exhaustive clause dispatchers (`execute_single_clause`,
`execute_mutable_bounded`) each gained one arm for the new variant, which
is the mechanical cost the Clause enum's own doc comment describes; their
per-function ceilings are refreshed, and the arms delegate to named
helpers so neither grew a body.
