+++ crates/kglite/src/graph/dir_graph/indexes.rs
@@ -1,0 +1,599 @@
+//! Secondary-index management on `DirGraph` — the hash equality, composite,
+//! and B-tree range index stores, plus the incremental maintenance the Cypher
+//! mutation executor calls after each write.
+//!
+//! Split out of `mod.rs` to keep it under the god-file LoC ceiling; index
+//! management is a self-contained concern with one entry point per index kind.
+//!
+//! The live stores are `#[serde(skip)]`; `populate_index_keys` snapshots their
+//! keys before save and `rebuild_indices_from_keys` replays them on load (both
+//! in `mod.rs`, with the other serialization helpers).
+
+use std::borrow::Cow;
+use std::collections::HashMap;
+
+use super::{DirGraph, IndexStats};
+use crate::datatypes::values::Value;
+use crate::graph::schema::{CompositeIndexKey, CompositeValue, IndexKey, InternedKey};
+use crate::graph::storage::backend::GraphBackend;
+use crate::graph::storage::GraphRead;
+use petgraph::graph::NodeIndex;
+
+impl DirGraph {
+    // ========================================================================
+    // Index Management Methods
+    // ========================================================================
+
+    /// Create an index on a property for a specific node type.
+    /// Returns the number of entries indexed.
+    ///
+    /// The id-alias / title-alias fields (e.g. `add_nodes(df, "Star",
+    /// "starId", "title")` makes `starId` the alias for the canonical
+    /// id) are intentionally NOT special-cased here: their indices
+    /// would build as empty (id/title live off the properties map).
+    /// Lookups against id-alias names route through `lookup_by_id_readonly`
+    /// in the matcher (`try_index_lookup`), which uses the auto-
+    /// maintained per-type `id_index` — no separate `create_index` call
+    /// required, and SET-on-id always stays in sync because id mutation
+    /// updates the id_index directly.
+    pub fn create_index(&mut self, node_type: &str, property: &str) -> usize {
+        // Store key uses the user's `property` name verbatim — the
+        // matcher's `try_index_lookup` indexes into `property_indices`
+        // by the unresolved user-facing key (matcher.rs:850), so the
+        // auto-maintenance path keeps things in sync only when the
+        // storage key matches.
+        let store_key = (node_type.to_string(), property.to_string());
+
+        // Mirror the matcher's property-READ path
+        // (`core/pattern_matching/matcher.rs::
+        // node_matches_properties_columnar`) so the index covers the
+        // same value-space MATCH consults. Three concerns the read
+        // path handles that `node.get_property()` did not:
+        //
+        // 1. **Alias resolution.** `starId` may be an id-alias for
+        //    `id`; same for title-aliases. Without resolving, we'd
+        //    look up "starId" in PropertyStorage and miss the data
+        //    (stored under "id").
+        // 2. **`id` / `title` are special.** Their values live in
+        //    `node_slots` (disk) / dedicated NodeData fields, NOT in
+        //    `properties`. The matcher reads them via `get_node_id`
+        //    / `get_node_title`; we do the same.
+        // 3. **Column-aware reads.** For mapped/disk graphs loaded
+        //    from .kgl, the actual values live in a `ColumnStore`,
+        //    not in the node's `PropertyStorage::Map`/`Compact`
+        //    snapshot. The backend's `get_node_property` knows how
+        //    to read from each storage type; `NodeData::get_property`
+        //    only reads the in-memory snapshot and silently returns
+        //    None for column-stored values.
+        let read_key = self.resolve_alias(node_type, property).to_string();
+        let interned_key = self.interner.get_or_intern(&read_key);
+        let mut index: HashMap<Value, Vec<NodeIndex>> = HashMap::new();
+
+        if let Some(node_indices) = self.type_indices.get(node_type) {
+            for idx in node_indices.iter() {
+                let value = if read_key == "id" {
+                    self.graph.get_node_id(idx)
+                } else if read_key == "title" {
+                    self.graph.get_node_title(idx)
+                } else {
+                    self.graph.get_node_property(idx, interned_key)
+                };
+                if let Some(value) = value {
+                    index.entry(value).or_default().push(idx);
+                }
+            }
+        }
+
+        let count = index.len();
+        self.property_indices.insert(store_key, index);
+        count
+    }
+
+    /// Install the single-property equality index for `(node_type, property)`,
+    /// **routing by storage backend**. Returns `(entries_indexed, persistent)`.
+    ///
+    /// On the Disk backend this builds the mmap-backed `PropertyIndex` and
+    /// reports `persistent = true`; the in-memory `property_indices` HashMap
+    /// would need multiple GB of heap for a ~13M-row type and be rebuilt on
+    /// every load. Memory and mapped backends build the HashMap via
+    /// [`Self::create_index`].
+    ///
+    /// **Every caller that installs an equality index on behalf of a user must
+    /// route through here.** [`Self::create_index`] is the in-memory primitive
+    /// and bypasses the disk decision; calling it directly on a disk graph is
+    /// the OOM this method exists to avoid. The counterpart
+    /// [`Self::drop_index`] already routes internally.
+    pub fn create_property_index_routed(
+        &mut self,
+        node_type: &str,
+        property: &str,
+    ) -> Result<(usize, bool), String> {
+        if let GraphBackend::Disk(disk) = &mut self.graph {
+            let count = disk
+                .build_property_index(node_type, property)
+                .map_err(|error| error.to_string())?;
+            return Ok((count, true));
+        }
+        Ok((self.create_index(node_type, property), false))
+    }
+
+    /// Drop an index on a property for a specific node type.
+    /// Returns true if the index existed and was removed.
+    pub fn drop_index(&mut self, node_type: &str, property: &str) -> Result<bool, String> {
+        if let GraphBackend::Disk(disk) = &mut self.graph {
+            return disk
+                .drop_property_index(node_type, property)
+                .map_err(|error| format!("persistent index removal failed: {error}"));
+        }
+        let key = (node_type.to_string(), property.to_string());
+        Ok(self.property_indices.remove(&key).is_some())
+    }
+
+    /// Check if an index exists for a given node type and property.
+    pub fn has_index(&self, node_type: &str, property: &str) -> bool {
+        let key = (node_type.to_string(), property.to_string());
+        self.property_indices.contains_key(&key)
+    }
+
+    /// Check if **any** index exists for `(node_type, property)` — the
+    /// in-memory `property_indices` HashMap *or* a persistent
+    /// disk-backed `PropertyIndex`. Used by `describe()` to annotate
+    /// schema output with `indexed=…` attributes so agents can tell
+    /// which properties hit an O(log N) path.
+    pub fn has_any_index(&self, node_type: &str, property: &str) -> bool {
+        if self.has_index(node_type, property) {
+            return true;
+        }
+        if let crate::graph::storage::backend::GraphBackend::Disk(dg) = &self.graph {
+            return dg.has_property_index(node_type, property);
+        }
+        false
+    }
+
+    /// Get all existing indexes as a list of (node_type, property) tuples.
+    pub fn list_indexes(&self) -> Vec<(String, String)> {
+        self.property_indices.keys().cloned().collect()
+    }
+
+    /// Look up nodes by property value using an index.
+    /// Returns None if no index exists, otherwise returns matching node indices.
+    pub fn lookup_by_index(
+        &self,
+        node_type: &str,
+        property: &str,
+        value: &Value,
+    ) -> Option<Vec<NodeIndex>> {
+        let key = (node_type.to_string(), property.to_string());
+        self.property_indices
+            .get(&key)
+            .and_then(|idx| idx.get(value))
+            .cloned()
+    }
+
+    /// Get statistics about an index.
+    pub fn get_index_stats(&self, node_type: &str, property: &str) -> Option<IndexStats> {
+        let key = (node_type.to_string(), property.to_string());
+        self.property_indices.get(&key).map(|idx| {
+            let total_entries: usize = idx.values().map(|v| v.len()).sum();
+            IndexStats {
+                unique_values: idx.len(),
+                total_entries,
+                avg_entries_per_value: if idx.is_empty() {
+                    0.0
+                } else {
+                    total_entries as f64 / idx.len() as f64
+                },
+            }
+        })
+    }
+
+    // ========================================================================
+    // Range Index Methods (B-Tree)
+    // ========================================================================
+
+    /// Create a range index (B-Tree) on a property for a specific node type.
+    /// Enables efficient range queries (>, >=, <, <=, BETWEEN).
+    /// Returns the number of unique values indexed.
+    pub fn create_range_index(&mut self, node_type: &str, property: &str) -> usize {
+        let key = (node_type.to_string(), property.to_string());
+        // Same alias-resolution + column-aware property read as
+        // `create_index` — see the comment there for the full
+        // rationale on why we don't use `node.get_property`.
+        let resolved = self.resolve_alias(node_type, property).to_string();
+        let interned_key = self.interner.get_or_intern(&resolved);
+        let mut index: std::collections::BTreeMap<Value, Vec<NodeIndex>> =
+            std::collections::BTreeMap::new();
+
+        if let Some(node_indices) = self.type_indices.get(node_type) {
+            for idx in node_indices.iter() {
+                let value = if resolved == "id" {
+                    self.graph.get_node_id(idx)
+                } else if resolved == "title" {
+                    self.graph.get_node_title(idx)
+                } else {
+                    self.graph.get_node_property(idx, interned_key)
+                };
+                if let Some(value) = value {
+                    index.entry(value).or_default().push(idx);
+                }
+            }
+        }
+
+        let count = index.len();
+        self.range_indices.insert(key, index);
+        count
+    }
+
+    /// Drop a range index. Returns true if it existed.
+    pub fn drop_range_index(&mut self, node_type: &str, property: &str) -> bool {
+        let key = (node_type.to_string(), property.to_string());
+        self.range_indices.remove(&key).is_some()
+    }
+
+    /// Range lookup: returns node indices where property value falls in the given range.
+    pub fn lookup_range(
+        &self,
+        node_type: &str,
+        property: &str,
+        lower: std::ops::Bound<&Value>,
+        upper: std::ops::Bound<&Value>,
+    ) -> Option<Vec<NodeIndex>> {
+        let key = (node_type.to_string(), property.to_string());
+        self.range_indices.get(&key).map(|btree| {
+            btree
+                .range((lower, upper))
+                .flat_map(|(_, indices)| indices.iter().copied())
+                .collect()
+        })
+    }
+
+    // ========================================================================
+    // Composite Index Methods
+    // ========================================================================
+
+    /// Create a composite index on multiple properties for a specific node type.
+    /// Composite indexes enable efficient lookups on multiple fields at once.
+    ///
+    /// Returns the number of unique value combinations indexed.
+    ///
+    /// Example: create_composite_index("Person", &["city", "age"]) allows efficient
+    /// queries like filter({'city': 'Oslo', 'age': 30}).
+    pub fn create_composite_index(&mut self, node_type: &str, properties: &[&str]) -> usize {
+        let key = (
+            node_type.to_string(),
+            properties.iter().map(|s| s.to_string()).collect(),
+        );
+
+        // Pre-resolve each property name (alias → canonical "id" /
+        // "title" when applicable) and pre-intern so the per-node
+        // loop is HashMap-only. Mirrors the matcher's read path —
+        // see `create_index`'s comment for the full rationale.
+        let resolved: Vec<String> = properties
+            .iter()
+            .map(|p| self.resolve_alias(node_type, p).to_string())
+            .collect();
+        let interned_keys: Vec<InternedKey> = resolved
+            .iter()
+            .map(|r| self.interner.get_or_intern(r))
+            .collect();
+
+        let mut index: HashMap<CompositeValue, Vec<NodeIndex>> = HashMap::new();
+
+        if let Some(node_indices) = self.type_indices.get(node_type) {
+            for idx in node_indices.iter() {
+                let values: Vec<Value> = resolved
+                    .iter()
+                    .zip(interned_keys.iter())
+                    .map(|(r, k)| {
+                        let v = if r == "id" {
+                            self.graph.get_node_id(idx)
+                        } else if r == "title" {
+                            self.graph.get_node_title(idx)
+                        } else {
+                            self.graph.get_node_property(idx, *k)
+                        };
+                        v.unwrap_or(Value::Null)
+                    })
+                    .collect();
+
+                // Only index if at least one value is non-null
+                if values.iter().any(|v| !matches!(v, Value::Null)) {
+                    index.entry(CompositeValue(values)).or_default().push(idx);
+                }
+            }
+        }
+
+        let count = index.len();
+        self.composite_indices.insert(key, index);
+        count
+    }
+
+    /// Drop a composite index.
+    /// Returns true if the index existed and was removed.
+    pub fn drop_composite_index(&mut self, node_type: &str, properties: &[String]) -> bool {
+        let key = (node_type.to_string(), properties.to_vec());
+        self.composite_indices.remove(&key).is_some()
+    }
+
+    /// Check if a composite index exists.
+    pub fn has_composite_index(&self, node_type: &str, properties: &[String]) -> bool {
+        let key = (node_type.to_string(), properties.to_vec());
+        self.composite_indices.contains_key(&key)
+    }
+
+    /// Get all existing composite indexes.
+    pub fn list_composite_indexes(&self) -> Vec<(String, Vec<String>)> {
+        self.composite_indices.keys().cloned().collect()
+    }
+
+    /// Look up nodes by composite values using a composite index.
+    /// Properties must match the order used when creating the index.
+    pub fn lookup_by_composite_index(
+        &self,
+        node_type: &str,
+        properties: &[String],
+        values: &[Value],
+    ) -> Option<Vec<NodeIndex>> {
+        let key = (node_type.to_string(), properties.to_vec());
+        let composite_value = CompositeValue(values.to_vec());
+
+        self.composite_indices
+            .get(&key)
+            .and_then(|idx| idx.get(&composite_value))
+            .cloned()
+    }
+
+    /// Get statistics about a composite index.
+    pub fn get_composite_index_stats(
+        &self,
+        node_type: &str,
+        properties: &[String],
+    ) -> Option<IndexStats> {
+        let key = (node_type.to_string(), properties.to_vec());
+        self.composite_indices.get(&key).map(|idx| {
+            let total_entries: usize = idx.values().map(|v| v.len()).sum();
+            IndexStats {
+                unique_values: idx.len(),
+                total_entries,
+                avg_entries_per_value: if idx.is_empty() {
+                    0.0
+                } else {
+                    total_entries as f64 / idx.len() as f64
+                },
+            }
+        })
+    }
+
+    /// Find a composite index that can be used for a given set of filter properties.
+    /// Returns the index key and whether all filter properties are covered.
+    pub fn find_matching_composite_index(
+        &self,
+        node_type: &str,
+        filter_properties: &[String],
+    ) -> Option<(CompositeIndexKey, bool)> {
+        // Sort filter properties for comparison
+        let mut sorted_filter: Vec<String> = filter_properties.to_vec();
+        sorted_filter.sort();
+
+        for key in self.composite_indices.keys() {
+            if key.0 == node_type {
+                let mut sorted_index: Vec<String> = key.1.clone();
+                sorted_index.sort();
+
+                // Check if index properties are a subset of or equal to filter properties
+                // For exact match, the index must cover exactly the filter fields
+                if sorted_index == sorted_filter {
+                    return Some((key.clone(), true)); // Exact match
+                }
+
+                // Check if index is a prefix of filter (can be used for partial filtering)
+                if sorted_filter.starts_with(&sorted_index)
+                    || sorted_index.iter().all(|p| sorted_filter.contains(p))
+                {
+                    return Some((key.clone(), false)); // Partial match
+                }
+            }
+        }
+        None
+    }
+
+    // ========================================================================
+    // Incremental Index Maintenance (called by Cypher mutations)
+    // ========================================================================
+
+    /// Update property, composite, and range indices after a new node is added.
+    /// Only updates indices that already exist for this node_type.
+    pub fn update_property_indices_for_add(&mut self, node_type: &str, node_idx: NodeIndex) {
+        // Collect single-property index updates (immutable borrow of self.graph)
+        let prop_updates: Vec<(IndexKey, Value)> = {
+            // Disk backend: node_weight materializes into the query arena,
+            // which must run under a DiskQueryGuard (arena protocol in
+            // disk/graph.rs). No-op on memory/mapped backends.
+            let _guard = self.graph.begin_query();
+            let node = match self.graph.node_weight(node_idx) {
+                Some(n) => n,
+                None => return,
+            };
+            self.property_indices
+                .keys()
+                .chain(self.range_indices.keys())
+                .filter(|(nt, _)| nt == node_type)
+                .filter_map(|key| {
+                    node.get_property(&key.1)
+                        .map(|v| (key.clone(), v.into_owned()))
+                })
+                .collect()
+        };
+        for (key, value) in &prop_updates {
+            if let Some(value_map) = self.property_indices.get_mut(key) {
+                value_map.entry(value.clone()).or_default().push(node_idx);
+            }
+            if let Some(btree) = self.range_indices.get_mut(key) {
+                btree.entry(value.clone()).or_default().push(node_idx);
+            }
+        }
+
+        // Collect composite index updates
+        let comp_updates: Vec<(CompositeIndexKey, CompositeValue)> = {
+            // Arena guard: see prop_updates block above.
+            let _guard = self.graph.begin_query();
+            let node = match self.graph.node_weight(node_idx) {
+                Some(n) => n,
+                None => return,
+            };
+            self.composite_indices
+                .keys()
+                .filter(|(nt, _)| nt == node_type)
+                .filter_map(|key| {
+                    let vals: Vec<Value> = key
+                        .1
+                        .iter()
+                        .map(|p| {
+                            node.get_property(p)
+                                .map(Cow::into_owned)
+                                .unwrap_or(Value::Null)
+                        })
+                        .collect();
+                    if vals.iter().any(|v| !matches!(v, Value::Null)) {
+                        Some((key.clone(), CompositeValue(vals)))
+                    } else {
+                        None
+                    }
+                })
+                .collect()
+        };
+        for (key, comp_val) in comp_updates {
+            if let Some(comp_map) = self.composite_indices.get_mut(&key) {
+                comp_map.entry(comp_val).or_default().push(node_idx);
+            }
+        }
+    }
+
+    /// Update property, range, and composite indices after a property value is changed.
+    /// Removes node from the old value bucket and adds to the new value bucket.
+    pub fn update_property_indices_for_set(
+        &mut self,
+        node_type: &str,
+        node_idx: NodeIndex,
+        property: &str,
+        old_value: Option<&Value>,
+        new_value: &Value,
+    ) {
+        let key = (node_type.to_string(), property.to_string());
+        // Update hash index
+        if let Some(value_map) = self.property_indices.get_mut(&key) {
+            if let Some(old_val) = old_value {
+                if let Some(indices) = value_map.get_mut(old_val) {
+                    indices.retain(|&idx| idx != node_idx);
+                    if indices.is_empty() {
+                        value_map.remove(old_val);
+                    }
+                }
+            }
+            value_map
+                .entry(new_value.clone())
+                .or_default()
+                .push(node_idx);
+        }
+        // Update range index
+        if let Some(btree) = self.range_indices.get_mut(&key) {
+            if let Some(old_val) = old_value {
+                if let Some(indices) = btree.get_mut(old_val) {
+                    indices.retain(|&idx| idx != node_idx);
+                    if indices.is_empty() {
+                        btree.remove(old_val);
+                    }
+                }
+            }
+            btree.entry(new_value.clone()).or_default().push(node_idx);
+        }
+
+        // Update any composite indices that include this property
+        self.update_composite_indices_for_property_change(node_type, node_idx, property);
+    }
+
+    /// Update property, range, and composite indices after a property is removed.
+    pub fn update_property_indices_for_remove(
+        &mut self,
+        node_type: &str,
+        node_idx: NodeIndex,
+        property: &str,
+        old_value: &Value,
+    ) {
+        let key = (node_type.to_string(), property.to_string());
+        if let Some(value_map) = self.property_indices.get_mut(&key) {
+            if let Some(indices) = value_map.get_mut(old_value) {
+                indices.retain(|&idx| idx != node_idx);
+                if indices.is_empty() {
+                    value_map.remove(old_value);
+                }
+            }
+        }
+        if let Some(btree) = self.range_indices.get_mut(&key) {
+            if let Some(indices) = btree.get_mut(old_value) {
+                indices.retain(|&idx| idx != node_idx);
+                if indices.is_empty() {
+                    btree.remove(old_value);
+                }
+            }
+        }
+
+        // Update any composite indices that include this property
+        self.update_composite_indices_for_property_change(node_type, node_idx, property);
+    }
+
+    /// Re-index a single node in all composite indices that include the changed property.
+    /// Reads current node properties to build the new composite value.
+    fn update_composite_indices_for_property_change(
+        &mut self,
+        node_type: &str,
+        node_idx: NodeIndex,
+        changed_property: &str,
+    ) {
+        let comp_keys: Vec<CompositeIndexKey> = self
+            .composite_indices
+            .keys()
+            .filter(|(nt, props)| nt == node_type && props.contains(&changed_property.to_string()))
+            .cloned()
+            .collect();
+
+        if comp_keys.is_empty() {
+            return;
+        }
+
+        // Read current node properties once. Arena guard: disk-backend
+        // node_weight materializes into the query arena (protocol in
+        // disk/graph.rs); no-op on memory/mapped backends.
+        let current_props: HashMap<String, Value> = {
+            let _guard = self.graph.begin_query();
+            match self.graph.node_weight(node_idx) {
+                Some(node) => node.properties_cloned(&self.interner),
+                None => return,
+            }
+        };
+
+        for key in comp_keys {
+            if let Some(comp_map) = self.composite_indices.get_mut(&key) {
+                // Remove node from all existing composite buckets
+                for indices in comp_map.values_mut() {
+                    indices.retain(|&idx| idx != node_idx);
+                }
+                // Remove empty buckets
+                comp_map.retain(|_, v| !v.is_empty());
+
+                // Build new composite value from current properties
+                let new_values: Vec<Value> = key
+                    .1
+                    .iter()
+                    .map(|p| current_props.get(p).cloned().unwrap_or(Value::Null))
+                    .collect();
+                if new_values.iter().any(|v| !matches!(v, Value::Null)) {
+                    comp_map
+                        .entry(CompositeValue(new_values))
+                        .or_default()
+                        .push(node_idx);
+                }
+            }
+        }
+    }
+}
